refactor: main 메소드에서의 게임 기능을 LadderGame 클래스 생성하여 책임 옮기기
