+++ src/main/java/ladder/domain/ladder/LadderFactory.java
@@ -1,7 +1,5 @@
 package ladder.domain.ladder;
 
-import ladder.domain.RandomLadder;
-
 public abstract class LadderFactory {
     public abstract RandomLadder createLadder(int height, int intervalCount);
 }
